// frontend/src/nobox-client/types.ts
export type StructureFieldType = typeof String | typeof Number | typeof Boolean | typeof Array;

export type CObject = Record<string, any>;

export type StructureItem<T = any, K = StructureFieldType> = {
  description?: string;
  comment?: string;
  required?: boolean;
  type: K;
  unique?: boolean;
  hashed?: boolean;
  defaultValue?: T;
};

export type Structure<T> = {
  [K in keyof T]: StructureItem<T[K]> | StructureFieldType;
};

export type SpaceAuthOptions = {
  active?: boolean;
  space: string;
  scope?: ('find' | 'insert' | 'delete' | 'update')[];
  token?: string;
};

export type SpaceFunctionValues<T> = {
  compulsoryParams: Partial<keyof T>[];
};

export type SpaceFunctionOptions<T> = {
  login?: SpaceFunctionValues<T>;
};

export type SpaceWebhooks = {
  onInsertUrl: string;
  onUpdateUrl: string;
};

export type Space<T> = {
  space: string;
  description: string;
  webhooks?: SpaceWebhooks;
  initialData?: T[];
  clear?: boolean;
  authOptions?: SpaceAuthOptions;
  structure: Structure<T>;
  functionOptions?: SpaceFunctionOptions<T>;
};